Tidy up and start pulling out People
