+++ src/submission/people.ts
@@ -1,0 +1,7 @@
+// TODO: shouldn't be here really.
+export interface Author {
+    firstName: string;
+    lastName: string;
+    email: string;
+    institution: string;
+}
